--- src/main-undertow/java/com/stubbornjava/examples/undertow/routing/ConstantStringHandler.java
@@ -1,22 +1,0 @@
-package com.stubbornjava.examples.undertow.routing;
-
-import io.undertow.server.HttpHandler;
-import io.undertow.server.HttpServerExchange;
-import io.undertow.util.Headers;
-
-// {{start:handler}}
-public class ConstantStringHandler implements HttpHandler {
-	private final String value;
-
-	public ConstantStringHandler(String value) {
-		this.value = value;
-	}
-
-	@Override
-	public void handleRequest(HttpServerExchange exchange) throws Exception {
-		exchange.getResponseHeaders().put(Headers.CONTENT_TYPE, "text/plain");
-		exchange.getResponseSender().send(value + "\n");
-	}
-
-}
-// {{end:handler}}
