--- src/main-undertow/java/com/stubbornjava/undertow/exchange/QueryParams.java
@@ -1,26 +1,0 @@
-package com.stubbornjava.undertow.exchange;
-
-import java.util.Deque;
-import java.util.Optional;
-
-import io.undertow.server.HttpServerExchange;
-
-public interface QueryParams {
-
-	default Optional<Long> queryParamAsLong(HttpServerExchange exchange, String name) {
-		return queryParam(exchange, name).map(Long::parseLong);
-	}
-
-	default Optional<String> queryParam(HttpServerExchange exchange, String name) {
-		return Optional.ofNullable(exchange.getQueryParameters().get(name))
-		               .map(Deque::getFirst);
-	}
-
-	default Optional<Integer> queryParamAsInteger(HttpServerExchange exchange, String name) {
-		return queryParam(exchange, name).map(Integer::parseInt);
-	}
-
-	default Optional<Boolean> queryParamAsBoolean(HttpServerExchange exchange, String name) {
-		return queryParam(exchange, name).map(Boolean::parseBoolean);
-	}
-}
